update logging and add noop mode

+++ lib/thor-tropo/uploader.rb
@@ -11,7 +11,6 @@ module ThorTropo
 
     no_commands {
 
-
       def initialize(opts={})
         @debug_mode   = opts[:debug_mode]
         @connection   = setup_connection(opts[:access_key],opts[:secret_key])
@@ -23,12 +22,13 @@ module ThorTropo
         local_file  = opts[:local_file]
         remote_file = opts[:path] + "/" + opts[:local_file].split("/")[-1]
         force       = opts[:force].nil? ? false : opts[:force]
+        noop        = opts[:noop].nil? ? false : opts[:noop]
 
         if file_exists? remote_file
 
           if force
 
-            file = send_file remote_file, local_file
+            send_file :remote_file => remote_file, :local_file => local_file, :noop => noop
 
           else
 
@@ -37,7 +37,7 @@ module ThorTropo
             if response.downcase.match /yes|ok|y/
               begin
 
-                file = send_file remote_file, local_file
+                send_file :remote_file => remote_file, :local_file => local_file, :noop => noop
 
               rescue Exception => e
                 say "[ TROPO ] Error: #{e}", :red
@@ -50,7 +50,7 @@ module ThorTropo
           end
         else
 
-          file = send_file remote_file, local_file
+          send_file :remote_file => remote_file, :local_file => local_file, :noop => noop
 
         end
       end
@@ -76,19 +76,29 @@ module ThorTropo
         !@bucket_name.files.head(key).nil?
       end
 
-      def send_file(remote_file,local_file)
+      def send_file(opts={})
+        local_file  = opts[:local_file]
+        remote_file = opts[:remote_file]
+
         say "[ TROPO ] - Uploading file.", :blue
-        file = @bucket_name.files.create(
-          :key    => remote_file,
-          :body   => File.open(local_file),
-          :public => true,
-          :content_type => get_mime_type(local_file),
-          :metadata => {
-            "x-amz-meta-sha256-hash" =>get_sha256_sum(local_file)
-          }
-        )
-        say "[ TROPO ] - Public URL: #{file.public_url}", :blue
-        return file
+        unless opts[:noop]
+          file = @bucket_name.files.create(
+            :key    => remote_file,
+            :body   => File.open(local_file),
+            :public => true,
+            :content_type => get_mime_type(local_file),
+            :metadata => {
+              "x-amz-meta-sha256-hash" =>get_sha256_sum(local_file)
+            }
+          )
+          public_url = file.public_url
+
+        else
+          public_url = " *** NOOP ** /#{remote_file}"
+
+        end
+          say "[ TROPO ] - Public URL: #{public_url}", :blue
+
       end
 
       def get_sha256_sum(file)
